chore: step3 피드백 반영
