+++ src/types/order.ts
@@ -1,0 +1,5 @@
+export interface Receiver {
+  name: string;
+  phone: string;
+  qty: number;
+}
